Bulk first commit for the moteist5, msp430f5xxx series and msp430x extended mode.

+++ sys/newlib_syscalls_default/syscalls.c
@@ -230,7 +230,7 @@ __attribute__((weak)) void heap_stats(void)
     }
 
     printf("heap: %ld (used %d, free %ld) [bytes]\n",
-           heap_size, minfo.uordblks, heap_size - minfo.uordblks);
+           heap_size, (int)minfo.uordblks, heap_size - minfo.uordblks);
 }
 #endif /* HAVE_HEAP_STATS */
